pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "minecraft-font-ocr"
version = "0.1.0"
description = "Purpose-built OCR engine for the Minecraft bitmap font"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.26",
    "opencv-python>=4.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["minecraft_ocr*"]
